requirements.txt
gitpython
python-dotenv
google-api-python-client
langchain_openai
langchain
huggingface_hub